[hive] MSVC compatibility fixes
MSVC doesn't allow anonymous struct types inside unions.
But Clang doesn't allow named struct types to be defined inside unions!
So, name the struct type *and* pull its definition outside of the union.
No intended functional change.

MSVC -W4 also warns about implicitly defaulted-as-deleted destructors;
so, explicitly delete these two destructors for clarity.

MSVC rejects a defaulted constructor for a type with a union involved
(they think it's defaulted-as-deleted). I think that's a bug, but it's
easy to fix. This constructor isn't trivial anyway, because of the
NSDMI on the `size_` member.

+++ test/hive_test.cpp
@@ -1606,9 +1606,9 @@ TYPED_TEST(hivet, EraseRandomlyUntilEmpty)
         for (int i = 0; i < 50 && !h.empty(); ++i) {
             auto it1 = h.begin();
             auto it2 = h.begin();
-            size_t n = h.size();
-            size_t offset = g() % (n + 1);
-            size_t len = g() % (n + 1 - offset);
+            int n = h.size();
+            int offset = g() % (n + 1);
+            int len = g() % (n + 1 - offset);
             std::advance(it1, offset);
             std::advance(it2, offset + len);
             EXPECT_DISTANCE(it1, it2, len);
@@ -1637,9 +1637,9 @@ TYPED_TEST(hivet, EraseInsertRandomly)
         for (int i = 0; i < 50 && !h.empty(); ++i) {
             auto it1 = h.begin();
             auto it2 = h.begin();
-            size_t n = h.size();
-            size_t offset = g() % (n + 1);
-            size_t len = g() % (n + 1 - offset);
+            int n = h.size();
+            int offset = g() % (n + 1);
+            int len = g() % (n + 1 - offset);
             std::advance(it1, offset);
             std::advance(it2, offset + len);
             EXPECT_DISTANCE(it1, it2, len);
